Added notes and hints spreadsheet
Added a Notes and Hints spreadsheet with notes about preparing the input file. Please let us keep expanding this list of 'lessons learnt' as we run into edge cases when debugging the model. The idea is for these notes to pertain to preparation of the input file to address these edge cases.